Make overcloud trust undercloud CA if undercloud has TLS
When the undercloud is enabled with TLS, the overcloud needs to trust
the CA, even if the overcloud is not using TLS.

The overcloud-ssl module already skips most of the process when
overcloud_ssl is false. So I removed the skipping of the overcloud-ssl
role from the playbooks in order for the CA injection template to be
generated.

Closes-Bug: #1731282
Depends-On: Ib88f6e4d561f9c8b5ba6215bbd9450a704b74eec
Change-Id: Iae6f1768018d37f898da1ad455475036896189c4