Make ovb-manage-stack role usable for upstream reproducers
Currently the ovb-manage-stack role does not have any way to
setup a non-root user and no way to dump an inventory file. This
works fine when used with the rest of quickstart, because we have
other code that does these things.

However in order to reproduce tripleo-ci OVB jobs we want to
simply provision an environment like tebroker does in those jobs.
Then we just hand that environment off to the tripleo-ci job
scripts.

This change adds two optional bits (both off by default) that will
add a non-root user and dump an inventory file.

Change-Id: I9ee4402572583d6acb7cfdbde79b4b2e41a53fdf