Allow using FQDN for ceph-mon setup
Currently, we have no way to tell ceph-ansible's ceph-mon role
that the nodes have a FQDN hostname. This is true on rdocloud.
This results in ceph-ansible waiting for a socket with the wrong
filename.

This patch adds the ability to pass the parameter ceph_use_fqdn,
which will result in adding the appropriate CephAnsibleExtraConfig
to the cloud-names environment.

Another patch (to tripleo-ci repo) will make this the default when
deploying on rdocloud.

Change-Id: I79cb56a10b4dd60c7bb7bab40bbdc5711f61f27b
Related-Bug: https://bugzilla.redhat.com/show_bug.cgi?id=1507888