Allow support for subunit2junitxml < 1.2
Versions of subunit2junitxml less than 1.2 do not support the
syntax we have in the validate tempest role. Namely:

  subunit2junitxml foo bar

Instead they need the destination file specified with '--output-to'

  subunit2junitxml foo --output-to=bar

As far as I can tell, the newer versions of subunit still support
the '--output-to' syntax. As such, we should use the syntax that
supports all versions.

Closes-Bug: 1720859

Change-Id: I3b52db71bb1b1b1f0faf241038348e6dedca943f