Add ui_validate_simple to the logs collected
The ui_validate_simple test has been failing of late and
it is hard to debug since logs are not collected
from localhost.
This review copies all the .sh and .log files
from the local working dir to the log directory so that
they are included in log collection.

Related-Bug: 1734928
Change-Id: I3bc646f5d1d9584cef4e624e5904f7e88c59442e