Fix TLS everywhere containerized deployments
This commit fixes the command line that gets created when a TLS
everywhere deployment is requested.
We pass to the deploy command line th file named
docker-services-tls-everywhere.yam but this should not be used anymore.
It doesn't exists anymore upstream and in addition if it's declared on
the command line AFTER the docker-ha.yml all the HA stuff breaks and the
deploy fails, because it tries to deploy overcloud HA services as docker
but they need instead pacemaker bundles.

Change-Id: I4116f44643a397d96115e74c4979e51c41677368