Calculated deploy timeout dynamically in CI
Calculated overcloud deploy timeout dynamically, basing on
passed ci_job_end_time variable in CI.
The difference between it and current time gives us the
time remaining in job. 10 minutes is left for logs collection
and 5 minutes is for tripleo-inventory to add overcloud nodes.
General timeout for playbook run remains as is.

Depends-On: I89b743fc16ea4c100ba21ddbe29081fa3e5479e9
Closes-Bug: #1738038

Change-Id: Icada6006377e0f27ea9a71ffad3320c99dd15e76