Add a hook to run a bash script at the end of the overcloud prep.
This would be useful for ffu mixed version upgrade testing.  The use
case is test environment only as we want to have a mixed version with
master-uc/netwon-oc.  This is not a normal environment use case and as
such we expect workaround to have this working.

The trouble of maintaining the extra workaround against having a not
too long feedback loop on the ffu upgrade seems worth the trouble.

Change-Id: Ib37b9f58e25b94fe7459b61355d5d15c3f0d36b8
Related-Bug: #1742924