Set SUPERMIN_* variables to working kernel
virt-customize(libguestfs) is not working with current
kernel 3.10.0-693.11.6.el7 in centos.
So until it's fixed in libguestfs or kernel, let's
use working kernel 3.10.0-693.el7.x86_64.
The kernel version is managed using following variable
which can be overridden when required.
Libguesfs requires following environment variables
to use the overridden kernel.

- SUPERMIN_KERNEL_VERSION: {{ libguestfs_kernel_override }}
- SUPERMIN_KERNEL: /boot/vmlinuz-{{ libguestfs_kernel_override }}
- SUPERMIN_MODULES: /lib/modules/{{ libguestfs_kernel_override }}

Change-Id: I129ec6c48d801cd605c4befda5ee00a025480413
Partial-Bug: #1743749