Upgrade UC and OC using tripleo-upgrade role.
Link tripleo-upgrade [0] role into oooq multinode playbook:

 - Added a containers preparation step for upgrades
 - Call tripleo-upgrade role with the specific parameters set in the
   job featureset.

[0] https://github.com/openstack/tripleo-upgrade

This keeps backward compatibility.

All the necessary bit for the docker switching repository
configuration will come in a later patch.  Here we just add the step.
The first target of this role being O->P, this isn't yet needed as
docker images are not upgrade in this scenario.

We are introducing a tech-debt here.  We have to refactor the set of
tasks introduced in the playbook to a proper role inside oooq-extra.
This should be done in a following refactor.

Related-Bug: #1749740

implements: blueprint tripleo-upgrade

Co-Authored-By: Sofer Athlan-Guyot <sathlang@redhat.com>

Change-Id: I92a1d915441686453ef0b241e4ba151c03595074