overcloud-deploy-post: detect keystone with undercloud's credentials
The current code path for detecting if keystone was deployed uses the
overcloud's credentials. This is not correct, as the output will
probably not be available there (unless you deployed a stack with a
similar output... but that would be quite a coincidence). So, this moves
the sourcing of the overcloud's credentials after getting the stack
output.

Change-Id: I5772dae90e7bfeba72440045ba3c772134ccc947