Fix overcloud image upload for containerized UC
Use overcloud image upload --http-boot only for containerized
undercloud.

Until we figure out a better way of managing http-boot argument when
preparing the image, let's add a conditional so we only use it for a
containerized undercloud and not for the OVB jobs which seems to not
work as of today.

Change-Id: I52cdcabc420276d05827e361d4189522fed57176
Related-Bug: #1748921
Closes-Bug: #1669186
Co-authored-by: Emilien Macchi <emilien@redhat.com>
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>