Make deploy_args_used fact cacheable.
Now that the CI job playbooks are ran
independently it has appeared undefined variables
which were being used among different playbooks.
This is the case of the "deploy_args_used" one,
which was being set in overcloud-deploy role to
be used in overcloud-upgrade.

This patch makes the fact cacheable, so
overcloud-upgrade role can make use of it.

Closes-Bug: #1750747

Change-Id: Iff90c3de2283e2711b49d91e23d8ed4a6d6888f3