Parametrize overcloud-prep-containers script name.
Upgrading from P->Q will imply running the
overcloud-prep-containers role twice. As the
name used in the script is fix, the second
time the role will run will overwrite the
previous script file, which could be useful
for debugging purposes.

This patch adds a new parameter
'overcloud_prep_containers_script' which sets
the name of the script used to configure the
docker containers.

Change-Id: I7952872b110b25f67d66de0c5ea4610bfb0f0cb1