Set network-feature-enabled.port_security to True
tempest.api.volume.test_volumes_snapshots.VolumesSnapshotTestJSON.
test_snapshot_create_offline_delete needs port_security feature.
In tempest, It is disabled by default.
We need to enable in order to run this tests and this test is default
in another featureset from Queens. So enabling this params for the
same.

Change-Id: I792c1372ee4261a85a4bcb05cb2c3db7dad313c9
Closes-Bug: #1753370