Modify OVB stack and resource identifiers to match te-broker
The ovb-manage-stack role creates OVB stacks in a similar way
to te-broker. However the patterns of naming the stacks and the
related resources were different. This review changes the naming
pattern of the ovb-manage-stack role to match that used with
te-broker to allow for a more uniform mangement of stacks
across tenants and host clouds.

Change-Id: I5c764b9eb9d49e75998d8fe39f41ac90ec00df73