Check repo setup script exists before content load.
In [0] there was an attempt to bypass the case
in which the repo setup script was not created
by using the Ansible lookup 'first_found'. However,
this lookup doesn't return the content of the file,
as the 'file' lookup does. It returns the name of
the file if the file exists.

This patch adds an Ansible step that checks the
existence of the repo setup script and loads the
content of the file if it exists, otherwise sets
the variable to false.

[0] https://review.openstack.org/#/c/549179/4

Change-Id: Iecf1534afff2bf3df4e2bfc23d0969fb57566b66
Closes-Bug: #1756037