Add option to specify the timeout for reproducer script
Current default timeout for jobs running in CI is 180
minutes which is not sufficient for some of the scenarios
to complete. This patch set's default timeout for reproducer
script to 240 minutes which should be sufficient for most
of the scenarios and add an option -t|--timeout to override
this default.

Change-Id: Iad864e2edb03d66b09d612178c2fbec0d4501fc1