Enable BuildRequires repository in DLRN for master
Since Rocky, RDO is spliting dependencies repo in two different
repos, one for runtime requirement with the URL used in previous
releases and a new one with the packages that are only required
to build packages. This repo is not needed to deploy RDO but it's
needed to build packages with DLRN.

This patch is enabling the BuildRequires repository in DLRN
buildroot when running it in TripleO-CI when the repo is
defined.

Depends-On: I74388434e3dccfea2754569a6bc568835d49b5e2
Change-Id: I4d14d1328578edf489f6560f369142cd845591aa