Don't use OS::Neutron::Port in subnode template
Using OS::Neutron::Port directly just to associate a floating IP is
strongly discouraged because the scheduling of the server becomes
constrained by the existing port.

Instead, let nova/neutron create the port implicitly and fetch the
resulting port ID to use for the floating IP association.

This change also adds a depends_on:router_interface to ensure the
interface exists before floating IPs attempt to use it.

This may improve the success rate of multinode jobs where nodes go to
ERROR due to schedule failures, but I'm not aware of any specific bug
which is tracking this.

Change-Id: Ie5cdd5a465040a0ece98a6103ad60b4a26610e42