Fixing tempest containerized script
Run tempest containerized creates a script, that requires the
public_net_id set inside the script, before it was being set in the
cleanup-network script, which doesn't export properly the variable,
causing error.
Also, since docker run as root, the  tempest_container.sh creates the
tempest files as root, so we need to use sudo to cleanup the
directories.

Change-Id: Ie5723df5e791dbe9572b7d3e5b7b71506563d11c