Fix changes regex
Previously if the number of patchsets was greater than 9 (meaning two
digits) the regex to try and capture the repo name was not properly
trimming all the numbers. Rather than requiring specific counts of
numbers, we should just trim any numbers from the ZUUL_CHANGES.

This could be seen by trying
ZUUL_CHANGES=openstack/tripleo-puppet-elements:master:refs/changes/72/521172/11
which would result in tripleo-puppet-elements1 which then would not be
matched by the tripleo-puppet-elements repo.

Change-Id: I4ad8b377c4b12a97d2d04859f5d05f8a8137be9d
Closes-Bug: #1766350