Prepare for DLRN modular build driver structure
Previously, the build-test-packages role modified DLRN's build_rpm.sh
script directly to make sure that the built rpms would not be installed
as part of the mock build in its chroot.

Since [1], DLRN will change the way it builds packages to use a driver
structure, and the build_rpm.sh script will no longer exist. On the flip
side, the default mock driver gives an option to skip installation in the
mock call, so we can simply use it.

This commit is compatible with the current DLRN version, since it checks
if the build_rpm.sh script is available and patches it if so, and
applies the project.ini template with the new option, which is ignored until
the DLRN patch is merged and released.

[1] - https://softwarefactory-project.io/r/12039

Change-Id: I0c45d764a86bf62da9e454af73cfa146beb45d0f