Setup iptables in nodepool-setup
The keystone-only deployment does not complete on our reproducer
environments setup using the nodepool-setup role. This is because
there is no default rule allowing the subnodes to communicate.
When iptables i brought up during deploy, we lose all network
connectivity on the vxlan tunnel as a result.

This patch resolves this issue by adding rules to the iptables
configuration which allow all traffic on the nodes private IPs.
The iptables setup during the deploy does not overwrite these rules,
so this allows the keystone-only deployment to succeed in our
reproducer environments.

Change-Id: If0af2608d1d07eb3931ce480501d0fa789ca4248