execute the build / install of zuul changes in upgrade
At the moment the build-test-packages and install-built-repo
are only called in the n version not the n+1.  If the
dependency is master it will not get installed, however
a queens patch would.

Once we flip the upgrade role we need to recheck
for zuul changes on the n+1 version and rebuild them
if they are found

Related-Bug: #1768997
Change-Id: I1d1401753f4576317805bc4c556b37f71faa60ed