Deployed Server - portmap should match neutron API
The cidr property for a Neutron subnet should contain
the network cidr, not the network prefixlen.

The example portmap for the fake neutron port in tripleo
depleoyed-server is not correct. We should pass the
netowork cidr in the subnets so that the fake neutron
port subnets attribute return something that match what
an actual neutron port would return.

Related-Bug: #1772396
Change-Id: I45f5cd1924c1d74cf12719dbed160ea960478850