Ensure IPA CA generates CRL on startup
If the CRL is not generated on startup, we could have deployments
that fail because the CRL is not present when puppet goes to get
it.  This fix will ensure the CRL is generated at the start to
avoid this issue.

See https://bugzilla.redhat.com/show_bug.cgi?id=1571897

Change-Id: Ia95f78257e23f36ce35b89e0a711b713e8d1b8b4