Add a variable to append paths in log collect list
This commit adds a new variable named artcl_collect_list_append that
gives the user the ability to append additional paths to what is the
default. So instead of redeclaring the (huge) amount of existing paths
with in addition the new ones, one can just add the ones he's interested
on inside the new variable.

Change-Id: I9843ce2a1159d550099f4d6dfe53dff22791a05a