Fix yum contentdir var for non altarch
Review https://review.openstack.org/#/c/568701/ created
contentdir file, but it's not working where file is not
present.
This patch fixes it for non altarch when file: contentdir
is not present on host.

Also change the fact name to yum_contentdir_data as
yum_contentdir conflicts with the when condition.

Change-Id: I99c095ab0da0cf969f0b7a1b3af890cc89c65454