Skip all non-kolla based images in containers-prepare-parameter.yaml
We're introducing a filter based on container image labels for the
tripleo-modify-image role that lets us skip non-kolla based images.
The remaining images are all good potential candidates for the update.

There is no need to maintain a list of excluded images anymore or to
push to the local registry in two separate steps.

Depends-On: Ie7f628bb54c83c78e810320c739afff158234804
Change-Id: Icad8918a7359a863b746b0bf9c2f2597a5c71bf8