Set update_repo to gating_repo_name parameter
In previous releases we used gating_repo_name parameter
to set the repositories to be used to update container
packages. After moving it to tripleo-modify-image role,
repo name is hardcoded to gating-repo but in some cases
we need to use a different name (as in RDO rdoinfo jobs),
so making it configurable using the same parameter seems
to be a good option.

Change-Id: I1a7a9edde53ecc85bb347f3c7dc3140aac06985e