Prevent mistral-db-manage not found error
The mistral DB manage populate code was to work around a specific
issue on Ocata. The bug is now closed as invalid and the workaround
causes the overcloud-prep-images.sh script to fail when
step_root_device_size is enabled, so remove the mistral DB population
call. I haven't observed any issue after removing it.

Change-Id: I29ee156494c381591f06a62ca6259506730c0b61
Related-Bug: #1649350