add delorean-current to repolist for updates
The list of containers that are updated are generated
from a list of rpms from the gating repo.  These
containers are updated w/ the gating repo and dlrn-current.

That makes the above set of containers out of sync w/
the rest of the containers.  The list of containers
that are updated needs to include changes required
by dlrn-current AND the gating repo.

The --enablerepo parameter for repoquery seems to support
comma-delimited lists, we'll take advantage of that so that we don't
need edit ansible-role-tripleo-modify-image parameter interface.

Co-Authored-By: Jiri Stransky <jistr@redhat.com>
Closes-Bug: #1786764
Change-Id: Ie12021ace7e9eb1695aa97ac5d97f3b948be9d86