Add ability to have dlrn build all the packages at once
By default, we currently loop through the packages to build and build
them one at a time. It is possible to have dlrn handle building all the
packages in a single invocation of dlrn. This patch adds a new variable
to the build-test-packages role called artg_build_one that switches
between these two package building methods.

Change-Id: I98fd6cde44f8e4b693d79fbd5abb77b8636a84c9