Rebase patch in reproducing as zuul does in CI
When we clone patches in CI with Zuul, it rebases them to latest
branch, so that we have always the latest code.
Reproducer should reproduce job exactly as it's in Ci, so let's
rebase patch when we clone it. It also ensures that we'll install
it because it'll have higher version that installed one and we'll
check the patch, not just latest master.

Change-Id: I3ee9382ccb85b5af06f1b0e1ddba1dc496cf04f4