Lint all the generated playbook after standalone deployment.
This task default to false, and should be changed in the featureset.

It ansible-lint all generated playbook and enable detection of error
in any playbook, even those that are not used in the
deployment (upgrade, update, ffu tasks for instances)

The current version only fails on syntax error (wrong ansible module
name mainly) as currently the generated playbooks misses name in a lot
of tasks.

When this is fixed we can enable error of any lint error.

Change-Id: Ibe0bd47cc88063f71c405b408342426888fbf88d