Revert "Revert "Add ability to have dlrn build all the packages at once""
This reverts commit 2af6e870456fa03d7acc5af3035eebaccb2e0388.

This includes a fix for the repo_built variable logic

It turns out ansible will always register a variable even if a task is
skipped. In this case, that means we are not properly collecting the
repo since repo_built is skipped in the second declaration.  This logic
works fine if artg_build_one is false but not if it is true (the
default)

Change-Id: I1408d9ded18cdacbd1fb5410e70de2ef8104f950