Allow pinning of ara in undercloud-setup
Newer versions of ara are causing newer versions of ansible to be pulled
in rather than using the packaged version provided. This change adds a
new variable called undercloud_ara_version which can be used to specify
a specific version for ara.  Currently this is set to 0.15.0 which works
with the ansible 2.4.4 version we currently have in some stable repos.

Change-Id: Id2899044b4b0ac19d587cfe7e321193d5f0c542d
Closes-Bug: #1795009