Remove usage of target_upgrade_version.
With the ongoing refactoring for upgrades jobs we
start passing the corresponding release file for
each playbook instead of passing the N_uc_N-1_oc.yaml
release file. One of the variables defined in the old
upgrades release file was target_upgrade_version, in
which it's specified the release we are upgrading to.
That's not needed anymore, as we can obtain that value
from the emit_releases.sh

Closes-Bug: #1812403
Change-Id: I5ac14e5fc28300e083fb4b921cbf4ba627904748