Switch the undercloud to deploy Podman by default starting from Stein
Starting from Stein, we want the Undercloud to be deploying Podman
containers instead of Docker containers.
This patch switches container_cli to be 'podman', and make sure we only
configure this new parameter in Stein and beyond.
Also run the container CLI with sudo, it's required to Podman on CentOS7.

blueprint podman-support
Change-Id: I21662de8ac0101a25c2b73a4650bb4af0f5df89a