OVB - external network masquerading
Change I4b956e8be92f1b7a71579d04c7e41c20da7ffdfa removes
the default masqurading in the masquerade-network
service environment.

Because OVB jobs use the undercloud as the router for
the external network, the undercloud must forward and
masquerade traffic on this network.

This change uses the hieradata override to add masquerading
for both the undercloud ctlplane network (192.168.24.0/24)
as well as the External (10.0.0.0/24) network.

Related-Bug: #1794729
Change-Id: I11b325458517334f97fc5f4754b4b39efff3a3f3