Set undercloud hostname from ansible variables
This sets the undercloud hostname from ansible variables instead of
having to call hostnamectl. This also sets a domain to the hostname if
needed.

The hostname is set with one of the following cases, in this order of
precedence

* When we have undercloud_cloud_domain defined, we use that for the FQDN
* When we have an FQDN we use that.
* If not, we default to <hostname>.tripleodomain.example.com

Note that these are only used if undercloud_undercloud_hostname wasn't
provided.

Depends-On: I7705744cab62f432e2e09bf1ee88bba809609cda 
Change-Id: I649d6cb87458f2bc9623c84830a45bdcf54432d4
Closes-Bug: #1801769