Harden yaml linting on modified files
Allows us to slowly adopt yamllint by having to fix only the files
we would be touching. Once we lower the amount of differences we
can switch it to run on all files.

In addition to having a more consistent code-style this solves
*recurring* bugs like duplicate dictionary keys, like the one
fixes in this change. Only few weeks ago we had to fix a similar bug.

Change-Id: I2ef275700b420a0f700613568e43576079cd4d03