Add ability to disable yum update with gating repo
Previously when the gating repo was enabled, we would blindly yum update
with the repo. This probably shouldn't be the case as we should just be
installing or updating the new packages after the repo has been
installed. However to maintain backwards compatibility, a new variable
called ib_gating_repo_update has been created with the default of True.
To disable the yum update after isntalling built repo, set this to
false.

Change-Id: I5172134738b5b3d5146e45f1e0a7720b6e1d5a2c
Closes-Bug: #1796639