Makes tempest sendmail failure easier to track
This task can fail with various errors and this change enables us
to spot its failure with a single LogStash query.

In order to avoid false positive string match when bash is using verbose
more we encoded the W char in the source, so a string match would happen
only when printed and not on the source code.

Change-Id: Icbb7f71aba2e9d2a4cf4e4c07f953fe3613e6707
Partial-Bug: 1806495