Add output for ip6tables and related ipv6 things
Until now, we don't have any output for iptables rules applied to v6.
It can be interesting to get them, especially since standard behavior
is to try first ipv6 if a name as associated v6. Even for localhost.

This patch also reformat a bit the output in the files, adding some
headers for a better reading.

Change-Id: I62e5c03fa38f5c4c266fbf27bd4f1ec0f3bf0633