Remove use of upgrade_init_command from multinode-overcloud-upgrade.yml.
The UpgradeInitCommand heat resource hasn't work for some releases, even
thought it has been recovered in [0] it added some complexity to the way
we were invoking tripleo-upgrade role, as we needed to load into the
upgrade_init_command variable the whole content of the script we wanted
to run to switch repos. For that reason it's better to relay on repo-setup
role to leave the repositories switched (without packages update) prior to
invoking tripleo-upgrade as it is done in the undercloud upgrade playbook.

Also, as tripleo-upgrade is now being imported by zuul, there is no need
to call dynamically to tripleo-upgrade as before, we can use include_role
instead.

Co-Authored By: Quique Llorente <ellorent@redhat.com>
[0] - 9164e6adbdce69a2b7c10e1aea6c5cb309b6754e

Change-Id: Ib35a971a4a4123933fb0d59c56655892fbd067a3
Closes-Bug: #1814531