Standardize os_tempest playbook and set vars dynamically
os_tempest role is now integrated with standalone job but the
problem is that we need to define lots of os_tempest vars, it
makes harder for end user to run the os_tempest playbook.
Moving the vars using set_fact and generating it dynamically
helps to reuse the same playbook in another playbook for
running tempest using os_tempest.

It also adds the default tests we want to run.

https://tree.taiga.io/project/tripleo-ci-board/task/794

Change-Id: If1d87bd122291c839dd2376dff3ac37d7784d5a3