Fix tempest container post logging
Currently when podman tempest container tempest tests fails, it
stores the exit status as fail and exit the run_tempest.sh script
leading to not copying the stuff from tempest container to host.

It decouples the step of copying and managing permission for tempest
post log processing and run only when tempest run finishes with in
a container.

https://tree.taiga.io/project/tripleo-ci-board/task/819

Change-Id: I9c7f97f3aa97c0ea061ba6aeb125501bf3a1adbf