Fix containerized tempest logic
We weren't testing containerized tempest in standalone job due the
missing logic on tempest-setup.sh script. This patch adds the logic for
when containerized tempest runs in standalone, copying the clouds.yaml
file instead of the stackrc/undercloudrc and also maintain the logic
when you execute tempest in the ovb, copying the rc file.
This also ensures that /var/lib/tempestdata and /var/log/tempest
directories exists in order to not fail the execution of containerized
tempest.

https://tree.taiga.io/project/tripleo-ci-board/issue/836

Depends-On: https://review.openstack.org/#/c/643212/
Change-Id: I90204150085a1c9b943f9e054c0420a8a5e66538
Related-Bug: 1819440