Changing the way you copy tempest data to shell
The copy module look for a directory or file locally, and copy to remote
machine. In upstream, this code works, because we run quickstart in the
undercloud machine. However, if you're running quickstart on a different
machine, it will try to search locally for /var/lib/tempestdata, which
doesn't exist. Use remote_src doesn't work for recursive directory. The
other option was to use synchronize but it's very slow for recursive
directories. The best option is to use command to copy locally.

https://tree.taiga.io/project/tripleo-ci-board/issue/836?kanban-status=2027734

Change-Id: I589241e657178909b115fcfdca907ac16dc79b71
Closes-Bug: 1819440