Add topology_map option to manage scale and flavors
The --*-flavor and --*-scale cli options have been deprecated for some
time. In order to remove them, we need to be able to pass
<rolename>Count and Overcloud<rolename>Flavor to the
deployment. This change adds a new configuration option which can be
used to manage these options in quickstart. In order to configure these
options as part of the deploy a user should specify the topology_map
with a dictionary using the rolename as a key with scale and flavor
underneath the flavor.

Example:

  topology_map:
    Controller:
      scale: 3
      flavor: controller
    Compute:
      scale: 1
      flavor: compute
    CephStorage:
      scale: 3
      flavor: ceph

NOTE: It is important that the rolenames in the key match the
capitalization of actual role.

Change-Id: Id2bc5692ebce3b2d937ca16665ef7af8390ded07