Allow to not set DockerInsecureRegistryAddress for standalone
Currently DockerInsecureRegistryAddress is always set
in standalone_parameters.yaml, this patch allows to not
set by setting role variable standalone_container_prep_updates
to false, which is default to true when standalone_upgrade=false.

This is required when want to build the containers on
host and use the same containers for the standalone
deployment without setting up an extra(one is already setup
by build-containers role) local registry.

Change-Id: Ic2bb4d79e5a73c183d209f51a626f875c06fb4a3