Rely on artg_rdo_packages to inject gating repo
Currently inject gating repo is triggered based on
repo_built.skipped. This is wrong as repo_built.skipped
is defined if the last project in the dependency change
is skipped.
This patch fixes it by relying on artg_rdo_packages
variable. So whenever atleast 1 rdo packge is built
inject repo will be triggered.

Related-Bug: #1815744
Change-Id: I3c363c82c12b356e9073289a6912edd878b0f6ab