Revert "Make git.openstack.org as default git server"
With the transition to opendev.org(no git.openstack.org now),
this change is no longer needed. The refs/changes are not
synched[1] to opendev.org so we can't use https://opendev.org
to Clone gated changes instead we need to use https://review.opendev.org.
Also this is a followup of [2].

[1] https://review.opendev.org/#/c/655238/3
[2] https://review.opendev.org/#/c/654606/

This reverts commit 74b1094b73884b4e1ba2b1e6f6204720d527b757.

Change-Id: Id9f4e7f3ef787d771ff51fa5d8ee7d8e11b1563b