undercloud-setup: Introduce tripleo_set_unique_hostname variable
In hostname playbook, add a task to modify /etc/hosts.
tripleo_set_unique_hostname is a new boolean which if set to true,
will add a line to /etc/hosts with an unique hostname
( {{ ansible_hostname }}-unique ) which will resolve the first IPv4
address found in the inventory.

It is useful to reproduce a deployment where the operator is using
hostnames to create SSL certificates and deploy the undecloud with the
public host being an hostname/fqdn and not an IP address.

Related-Bug: #1763776
Change-Id: I6425499e5d248e6cb2285f14dd9c6e716122b90b