Download and define base image if configured
When we build the image we need a base image to be defined.
For CentOS it's done automatically in client, but if we need to
build something else we need to download it and define it
explicitly.
base_image_url - URL of base image
base_image_path - path to save it on host

Change-Id: I10bd9868fcfd3a4d5992fc7d7b6200e908e75e25