Move os_tempest role var at the top of play level
Once tempest run finishes using os_tempest role,
python_venv_build noop handler[1.] is called and the playbook
fails with tempest_install_method undefined var as it's scope
is only for that task when role is called.

Moving the var at the top of the play keeps the scope of the
var for the whole play.

Closes-Bug: #https://bugs.launchpad.net/tripleo/+bug/1838496

Change-Id: I58082b307cb5b3f9ff908c29c8e24b9b970561de
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>