Add basic minion validation
The minion doesn't actually have any api services so we validate it by
checking the service status on the undercloud. This change adds a role
to validate the heat engine and ironic conductor status on the
undercloud includes additional hosts (undercloud + minion) and there are
no failed services.

Change-Id: I41859f93980dd8cd723ea4ad0f4e6592a78d0082
Related-Blueprint: undercloud-minion