Increase size of cinder LVM backend
Increase the size of the cinder LVM backend to 20GB, and re-enable
tempest tests that were previously skipped due to space limitation.

Tempest's volume tests don't actually need 20GB of storage, but rapidly
creating and deleting volumes in a short time span may cause cinder's
scheduler to report a reduced amount of free space. The actual amount
of space in use is updated by the cinder-volume service, but this is
done periodically. Increasing the size of the LVM backend allows the
scheduler to sustain a burst of volume creation requests during the
time interval between updates from the volume service on the actual
amount of storage in use.

Closes-Bug: #1841071
Change-Id: Ie02211f3a6706f6fa1102fb8dc5ee0e9e54252cc