[reproducer]Set registry_login_enabled to false
Since in reproducer, there is no access to RDO registry and
in our jobs registry_login_enabled: true is enabled by
default [1.] and it tries to authenticate with the registry and
we have no credentials leading to failure in the base job itself.

Setting it to false fixes the issue.

[1.] https://review.rdoproject.org/r/#/c/22304/

Change-Id: I102683b637d60afbdb3898c811ba5961b69fede4
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>