Add container-update role and containers prepare step for Undercloud upgrades.
Since Rocky, the Undercloud is also containerized. The way to
tell the undercloud which containers to use is by setting the
right registry in containers-prepare-parameters.yaml file.

This patch adds a new role named container-update which makes use
of the parameters set in container-prep and substitute the relevant
parameters inside containers-prepare-parameter.yaml file. The input
for the role is the file used to configure the ContainerImagePrepare
environment variable (defaults to containers-prepare-parameter.yaml).

As part of the patch, it's been added a call to the newly added repo
container-update into multinode-undercloud-upgrade.yaml playbook to
update all needed parameters in containers-prepare-parameter.yaml
before the Undercloud's upgrade.

Change-Id: Icd4553f7c236080ebc529079142110a0b2e0ac2d