Add containerized_undercloud_upgrade variable.
When upgrading from Queens to Rocky, we are jumping from a
non-containerized undercloud into a containerized one.
However, this scenario hasn't been considered yet. When
submitting [0] we add the logic to upgrade always to a
containerized undercloud, we need to have a variable that
allows us differentiate between the type (containerized or
non-containrized) of undercloud we deploy and upgrade to.

[0] - Ibcdfd71b9280615aedc4d59ef0c566c475f15d0b

Related-Bug: #1855856

Change-Id: I30b4789adcf348c636d3679523075b22f8956ff1