Render containerized-prepare-parameters.yaml during upgrade if not found.
Generally, if we upgrade from a containerized undercloud the
containers-prepare-paramters.yaml file will be already present (as
it is being rendered during the undercloud-deploy). However, if we
upgrade the Undercloud from Queens (non-containerized) to Rocky
(containerized) the file won't be present. We need to render it
during the container-update role so we can set the right values.

Change-Id: I3fd411ef753bdd1f4ce5d6befa726d82e9073458
Related-Bug: #1855856