Use osp_release flag to set selinux mode
In Upstream and rdo-cloud tripleo ci jobs on RHEL & CentOS, we
use selinux mode to permissive but currently it is harded for
CentOS only.

In Downstream jobs, we use enforcing mode. So instead of depending
upon ansible_distribution, we can rely on osp_release to toggle
selinux mode and will work for both centOS and RHEL.

Related-Bug: 1853028

Change-Id: I6a6449777ea28198002b8c028a345ab16b733901
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>