Override tempest nova scheduler filters
In Rocky and Queens tempest contains a default value for nova scheduler
filters equal to all, which results in the mentioned bug when tests which
are meant to be executed against specific filters are not enabled in the
system.

Overriding filters in tempest to only those which are available in the system
will skip the tests which are meant for different filters.

Change-Id: I98532113d3ad1339fdc4869686ad875b304f094e
Closes-bug: 1857365