Fix project name mapping to DLRN project name (v2)
This fixes failures produced when project name
doesn't starts with 'openstack/'

This is intended to unblock the use of tripleo-ci
that is now failing with projects that are not
hosted under the 'openstack/' prefix (like for
example x/tobiko).

Note: removed set -u because this can break source activate on old
versions of virtualenv.

The role should not fail when it fails to find a mapping, so it was
refactored to assured that.

Also assures that the html report of Ansible is colored, making it
easier to read.

Change-Id: Ibb562f818419373700177f3b2e665cd4eb5e32ef
Partial-Bug: #1859167
Partial-Bug: #1858890